requirements.txt
opencv_python_headless
requests==2.28.1
lxml==4.9.1
dataclasses
Pillow==7.2.0
discord.py==1.7.3